feat(API): Routes
Seperate the route between the standard one and the apiRoutes
